pre-update for v1.5

+++ src/ModioX/Extensions/FtpExtensions.cs
@@ -14,13 +14,15 @@ namespace ModioX.Extensions
     internal static class FtpExtensions
     {
         /// <summary>
-        ///     Upload a local file to the specified location on the console
+        /// Upload a local file to the specified location on the console
         /// </summary>
         /// <param name="ftpConnection">PS3 IP address</param>
         /// <param name="localFile">Path of the local file</param>
         /// <param name="consoleFile">Path of the uploading file directory</param>
-        internal static void UploadFile(FtpConnection ftpConnection, string localFile, string consoleFile)
+        internal static void UploadFile(string localFile, string consoleFile)
         {
+            FtpConnection ftpConnection = MainWindow.FtpConnection;
+
             string dirPath = consoleFile.Contains("/")
                     ? consoleFile.Substring(0, consoleFile.LastIndexOf('/')) + '/'
                     : "dev_hdd0/";
@@ -39,20 +41,17 @@ namespace ModioX.Extensions
         }
 
         /// <summary>
-        ///     Uninstall a specified file from the console
+        /// Uninstall a specified file from the console
         /// </summary>
         /// <param name="ftpConnection">Mod to uninstall</param>
         /// <param name="consoleFile">Mod to uninstall</param>
         internal static void DeleteFile(FtpClient ftpClient, string consoleFile)
         {
-            if (ftpClient.FileExists(consoleFile))
-            {
-                ftpClient.DeleteFile(consoleFile);
-            }
+            if (ftpClient.FileExists(consoleFile)) ftpClient.DeleteFile(consoleFile);
         }
 
         /// <summary>
-        ///     Uninstall a specified file from the console
+        /// Uninstall a specified file from the console
         /// </summary>
         /// <param name="ftpConnection">Mod to uninstall</param>
         /// <param name="consolePath">Mod to uninstall</param>
@@ -64,14 +63,11 @@ namespace ModioX.Extensions
 
             ftpClient.SetWorkingDirectory(dirPath);
 
-            if (ftpClient.DirectoryExists(consolePath))
-            {
-                ftpClient.DeleteDirectory(consolePath);
-            }
+            if (ftpClient.DirectoryExists(consolePath)) ftpClient.DeleteDirectory(consolePath);
         }
 
         /// <summary>
-        ///     Downloads the specified console file to the computer
+        /// Downloads the specified console file to the computer
         /// </summary>
         /// <param name="localFile">Path of the local file</param>
         /// <param name="consoleFile">Path of the uploading file directory</param>
@@ -85,7 +81,16 @@ namespace ModioX.Extensions
 
             ftpConnection.SetLocalDirectory(Path.GetDirectoryName(localFile));
             ftpConnection.SetCurrentDirectory(dirPath);
-            ftpConnection.GetFile(consoleFile, localFile, false);
+
+            if (File.Exists(localFile))
+            {
+                File.Delete(localFile);
+            }
+
+            if (FileExists(consoleFile))
+            {
+                ftpConnection.GetFile(consoleFile, localFile, false);
+            }
         }
 
         internal static bool IsDirectoryEmpty(FtpClient ftpClient, string consolePath)
@@ -113,7 +118,7 @@ namespace ModioX.Extensions
         }
 
         /// <summary>
-        ///     Downloads the specified console file to the computer
+        /// Downloads the specified console file to the computer
         /// </summary>
         /// <param name="filePath">Path of the uploading file directory</param>
         internal static bool FileExists(string filePath)
@@ -130,7 +135,7 @@ namespace ModioX.Extensions
         }
 
         /// <summary>
-        ///     Renames the specified file to a new name
+        /// Renames the specified file to a new name
         /// </summary>
         /// <param name="ftpConnection"></param>
         /// <param name="filePath"></param>
@@ -138,15 +143,15 @@ namespace ModioX.Extensions
         internal static void RenameFileOrFolder(FtpConnection ftpConnection, string filePath, string newName)
         {
             // Use the WebRequest method because FtpConnection throws an error and this doesn't, strange
-            FtpWebRequest request = (FtpWebRequest) WebRequest.Create("ftp://" + ftpConnection.Host + filePath);
+            FtpWebRequest request = (FtpWebRequest)WebRequest.Create("ftp://" + ftpConnection.Host + filePath);
             request.Method = WebRequestMethods.Ftp.Rename;
             request.Credentials = new NetworkCredential(MainWindow.ConsoleProfile.Username, MainWindow.ConsoleProfile.Password);
             request.RenameTo = newName;
-            _ = request.GetResponse();
+            request.GetResponse();
         }
 
         /// <summary>
-        ///     Downloads the specified console file to the computer
+        /// Creates the specified directory path on the console.
         /// </summary>
         /// <param name="consolePath"></param>
         /// <returns>Whether the directory was created or not.</returns>
@@ -155,47 +160,52 @@ namespace ModioX.Extensions
             // Use the WebRequest method because FtpConnection throws an error and this doesn't, strange
             WebRequest request = WebRequest.Create("ftp://" + MainWindow.ConsoleProfile.Address + consolePath);
             request.Method = WebRequestMethods.Ftp.MakeDirectory;
-            request.Credentials =
-                new NetworkCredential(MainWindow.ConsoleProfile.Username, MainWindow.ConsoleProfile.Password);
+            request.Credentials = new NetworkCredential(MainWindow.ConsoleProfile.Username, MainWindow.ConsoleProfile.Password);
 
-            using (FtpWebResponse resp = (FtpWebResponse) request.GetResponse())
+            using FtpWebResponse resp = (FtpWebResponse)request.GetResponse();
+            return resp.StatusCode == FtpStatusCode.PathnameCreated;
+        }
+
+        internal static void CreateDirectories(string consolePath)
+        {
+            string[] folderArray = consolePath.Split('/');
+            string folderName = "";
+            for (int i = 0; i < folderArray.Length; i++)
             {
-                if (resp.StatusCode == FtpStatusCode.PathnameCreated)
+                if (!string.IsNullOrEmpty(folderArray[i]))
                 {
-                    return true;
+                    folderName = string.IsNullOrEmpty(folderName) ? folderArray[i] : folderName + "/" + folderArray[i];
+                    WebRequest request = WebRequest.Create("ftp://" + MainWindow.ConsoleProfile.Address + "/" + folderName);
+                    request.Method = WebRequestMethods.Ftp.MakeDirectory;
+                    request.Credentials = new NetworkCredential(MainWindow.ConsoleProfile.Username, MainWindow.ConsoleProfile.Password);
+                    //var response = request.GetResponse();
+                    //var test = response.ToString();
                 }
-
-                return false;
             }
         }
 
         internal static void DownloadDirectory(string consolePath, string localPath)
         {
             string url = "ftp://" + MainWindow.ConsoleProfile.Address + consolePath;
-            NetworkCredential credentials =
-                new NetworkCredential(MainWindow.ConsoleProfile.Username, MainWindow.ConsoleProfile.Password);
+            NetworkCredential credentials = new NetworkCredential(MainWindow.ConsoleProfile.Username, MainWindow.ConsoleProfile.Password);
 
-            FtpWebRequest request = (FtpWebRequest) WebRequest.Create(url);
+            FtpWebRequest request = (FtpWebRequest)WebRequest.Create(url);
             request.Method = WebRequestMethods.Ftp.ListDirectoryDetails;
             request.Credentials = credentials;
             request.Timeout = -1;
 
             List<string> lines = new List<string>();
 
-            using (FtpWebResponse listResponse = (FtpWebResponse) request.GetResponse())
+            using (FtpWebResponse listResponse = (FtpWebResponse)request.GetResponse())
             using (Stream listStream = listResponse.GetResponseStream())
             using (StreamReader listReader = new StreamReader(listStream))
             {
-                while (!listReader.EndOfStream)
-                {
-                    lines.Add(listReader.ReadLine());
-                }
+                while (!listReader.EndOfStream) lines.Add(listReader.ReadLine());
             }
 
             foreach (string line in lines)
             {
-                string[] tokens =
-                    line.Split(new[] {' '}, 9, StringSplitOptions.RemoveEmptyEntries);
+                string[] tokens = line.Split(new[] { ' ' }, 9, StringSplitOptions.RemoveEmptyEntries); 
                 string name = tokens[8];
                 string permissions = tokens[0];
 
@@ -204,37 +214,32 @@ namespace ModioX.Extensions
 
                 if (permissions[0] == 'd')
                 {
-                    if (!Directory.Exists(localFilePath))
-                    {
-                        _ = Directory.CreateDirectory(localFilePath);
-                    }
+                    if (!Directory.Exists(localFilePath)) Directory.CreateDirectory(localFilePath);
 
                     DownloadDirectory(fileUrl, localFilePath);
                 }
                 else
                 {
-                    FtpWebRequest downloadRequest = (FtpWebRequest) WebRequest.Create(fileUrl);
+                    FtpWebRequest downloadRequest = (FtpWebRequest)WebRequest.Create(fileUrl);
                     downloadRequest.Method = WebRequestMethods.Ftp.DownloadFile;
                     downloadRequest.Credentials = credentials;
                     downloadRequest.Timeout = -1;
 
-                    using (FtpWebResponse downloadResponse = (FtpWebResponse) downloadRequest.GetResponse())
-                    using (Stream sourceStream = downloadResponse.GetResponseStream())
-                    using (Stream targetStream = File.Create(localFilePath))
+                    using FtpWebResponse downloadResponse = (FtpWebResponse)downloadRequest.GetResponse();
+                    using Stream sourceStream = downloadResponse.GetResponseStream();
+                    using Stream targetStream = File.Create(localFilePath);
+                    byte[] buffer = new byte[10240];
+                    int read;
+                    while (sourceStream != null && (read = sourceStream.Read(buffer, 0, buffer.Length)) > 0)
                     {
-                        byte[] buffer = new byte[10240];
-                        int read;
-                        while (sourceStream != null && (read = sourceStream.Read(buffer, 0, buffer.Length)) > 0)
-                        {
-                            targetStream.Write(buffer, 0, read);
-                        }
+                        targetStream.Write(buffer, 0, read);
                     }
                 }
             }
         }
 
         /// <summary>
-        ///     Gets the folder names inside the specified console path
+        /// Gets the folder names inside the specified console path
         /// </summary>
         /// <param name="consolePath">Path of the uploading file directory</param>
         internal static List<string> GetFolderNames(string consolePath)
@@ -258,33 +263,30 @@ namespace ModioX.Extensions
         }
 
         /// <summary>
-        ///     Prompts the user to choose their user id
+        /// Prompts the user to choose their user id
         /// </summary>
         /// <param name="owner">Console IP Address</param>
         /// <returns></returns>
-        public static string GetUserId(Form owner)
+        public static string GetUserId()
         {
             List<string> userIds = GetFolderNames("/dev_hdd0/home/");
 
-            List<string> userNames = (from userId in userIds
+            List<string> userNames = (from userId in userIds 
                                       select $"{userId} ({GetUserNameFromUserId(userId)})").ToList();
 
             if (userIds.Count > 0)
             {
-                string userId = DialogExtensions.ShowListInputDialog(owner, "User Profile IDs", userNames).Split()[0];
+                string userId = DialogExtensions.ShowListInputDialog("User Profile IDs", userNames).Split()[0];
 
-                if (userId != null)
-                {
-                    return userId;
-                }
+                if (userId != null) return userId;
             }
 
-            DarkMessageBox.Show(owner, "Could not find any users. Make sure you have created at least one user profile.", "No Users Found", MessageBoxIcon.Error);
+            DarkMessageBox.ShowError("Could not find any users. Make sure you have created at least one user profile.", "No Users Found");
             return null;
         }
 
         /// <summary>
-        ///     Prompts the user to choose their user id
+        /// Prompts the user to choose their user id
         /// </summary>
         /// <param name="userId">Console IP Address</param>
         /// <returns></returns>
@@ -294,15 +296,13 @@ namespace ModioX.Extensions
 
             string usernameFile = $"/dev_hdd0/home/{userId}/localusername";
 
-            using (Stream stream = ftpClient.OpenRead(usernameFile))
-            using (StreamReader streamReader = new StreamReader(stream))
-            {
-                return streamReader.ReadToEnd();
-            }
+            using Stream stream = ftpClient.OpenRead(usernameFile);
+            using StreamReader streamReader = new StreamReader(stream);
+            return streamReader.ReadToEnd();
         }
 
         /// <summary>
-        ///     Returns the USB path if one is connected to the console port
+        /// Returns the USB path if one is connected to the console port
         /// </summary>
         /// <returns></returns>
         public static string GetUsbPath()
@@ -325,7 +325,7 @@ namespace ModioX.Extensions
                 }
             }
 
-            _ = DarkMessageBox.Show(MainWindow.Window, "You do not have a USB device connected to your console.", "Error", MessageBoxIcon.Error);
+            DarkMessageBox.ShowError("You do not have a USB device connected to your console.", "Error");
             return null;
         }
     }
